+++ lib/src/stdlib/collection.dart
@@ -11,6 +11,7 @@ import 'package:d4rt/src/stdlib/collection/splay_tree_set.dart';
 import 'package:d4rt/src/stdlib/collection/double_linked_queue.dart';
 import 'package:d4rt/src/stdlib/collection/unmodifiable_map_view.dart';
 import 'package:d4rt/src/stdlib/collection/unmodifiable_set_view.dart';
+import 'package:d4rt/src/stdlib/collection/map_view.dart';
 
 export 'package:d4rt/src/environment.dart';
 export 'package:d4rt/src/stdlib/collection/hash_map.dart';
@@ -25,6 +26,7 @@ export 'package:d4rt/src/stdlib/collection/splay_tree_set.dart';
 export 'package:d4rt/src/stdlib/collection/double_linked_queue.dart';
 export 'package:d4rt/src/stdlib/collection/unmodifiable_map_view.dart';
 export 'package:d4rt/src/stdlib/collection/unmodifiable_set_view.dart';
+export 'package:d4rt/src/stdlib/collection/map_view.dart';
 
 class CollectionStdlib {
   static void register(Environment environment) {
@@ -41,5 +43,6 @@ class CollectionStdlib {
     environment.defineBridge(DoubleLinkedQueueCollection.definition);
     environment.defineBridge(UnmodifiableMapViewCollection.definition);
     environment.defineBridge(UnmodifiableSetViewCollection.definition);
+    environment.defineBridge(MapViewCollection.definition);
   }
 }