+++ lib/src/stdlib/core/object.dart
@@ -5,12 +5,31 @@ class ObjectCore {
         nativeType: Object,
         name: 'Object',
         typeParameterCount: 0,
+        constructors: {
+          '': (visitor, positionalArgs, namedArgs) => Object(),
+        },
+        staticMethods: {
+          'hash': (visitor, positionalArgs, namedArgs) {
+            return Object.hashAll(positionalArgs);
+          },
+          'hashAll': (visitor, positionalArgs, namedArgs) {
+            final iterable = positionalArgs[0] as Iterable;
+            return Object.hashAll(iterable);
+          },
+          'hashAllUnordered': (visitor, positionalArgs, namedArgs) {
+            final iterable = positionalArgs[0] as Iterable;
+            return Object.hashAllUnordered(iterable);
+          },
+        },
         methods: {
           '==': (visitor, target, positionalArgs, namedArgs) {
             return target == positionalArgs[0];
           },
           'toString': (visitor, target, positionalArgs, namedArgs) {
             return target.toString();
+          },
+          'noSuchMethod': (visitor, target, positionalArgs, namedArgs) {
+            return (target as dynamic).noSuchMethod(positionalArgs[0] as Invocation);
           },
         },
         getters: {